requirements.txt
streamlit==1.40.2
plotly==5.24.1
pandas==2.2.3
pydantic==2.10.3
pyyaml==6.0.2
requests==2.32.3

# Dev / Test
pytest==8.3.4
black==24.10.0
isort==5.13.2
flake8==7.1.1
